+++ main.py
@@ -1,11 +1,18 @@
-# standart library imports
 import logging
+import time
 import traceback
-# third-party imports
 
-# inside imports
 from logs import logging_setup, set_logging
-from utils import system_setup, import_plugins, find_plugins
+from utils import system_setup, import_plugins, find_plugins, admin
+
+log = logging.getLogger("main")
+
+if admin():
+    log.error("Program should not be run with super user (sudo or admin) privileges. Exiting...")
+    sys.exit()
+
+start_time = time.time()
+
 from app import App
 from api import app as app_api
 
@@ -13,14 +20,15 @@ from data.constants import PLUGINS_FOLDER
 # from gui import main
 
 
-log = logging.getLogger("main")
-
-
 def main():
     # noinspection PyBroadException
     try:
+        # system preparations
+        log.debug("started running...")
+
         system_setup()
         set_logging(True)
+
         app = App(app_api)
 
         # plugins
@@ -29,7 +37,7 @@ def main():
 
         # main app and api initialization
         app.initialize()
-        app.start()
+        app.start(start_time)
 
     except Exception as e:
         log.debug(f"App loop ended with the following error: {e}: \n{traceback.format_exc()} ")